+++ lib/action_cable/subscription_adapter/anycable.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require_relative "any_cable"
+
+module ActionCable
+  module SubscriptionAdapter
+    # Aliasing adapter
+    Anycable = AnyCable
+  end
+end